LayoutTests/http/tests/websocket/tests/hybi/binary-type-expected.txt

Test WebSocket.binaryType attribute.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS ws.binaryType is "blob"
PASS ws.binaryType is "arraybuffer"
PASS ws.binaryType is "blob"
PASS ws.binaryType is "blob"
PASS ws.binaryType is "blob"
PASS ws.binaryType is "blob"
PASS successfullyParsed is true

TEST COMPLETE
